# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "back-end part for mobile application.",
        "title": "Learn-Share API",
        "contact": {
            "name": "Ruslan's Support",
            "url": "https://t.me/Ruslan20007",
            "email": "ruslanrbb8@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:81",
    "basePath": "/api",
    "paths": {
...
        "/teacher": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all info about teacher (user info + teacher + his skills) by user id in token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get teacher data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_teacher.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user id by jwt token, and he became teach (if he was not be registrate himself as teacher)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "User registrate also as teacher",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/teacher/schedule": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get lessons times from teacher schedule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get times from schedule",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_times.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add time to teacher schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Add time to schedule",
                "parameters": [
                    {
                        "description": "datetime",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/add_time.request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/teachers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get full teachers data (their user data, teacher data and skills)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get full teachers data",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter my teachers",
                        "name": "is_mine",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_teachers.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
        "/teachers/{id}": {
            "get": {
                "description": "Get all info about teacher (user info + teacher + his skills) by his UserID in route (/api/teachers/{id})",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get teacher data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher's UserID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_teacher.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/teachers/{id}/schedule": {
            "get": {
                "description": "Get lessons times from teacher schedule (by his UserID)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get times from schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher's UserID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_times.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "add_time.request": {
            "type": "object",
            "required": [
                "datetime"
            ],
            "properties": {
                "datetime": {
                    "type": "string",
                    "example": "2025-02-01T00:00:00Z"
                }
            }
        },
...
        "get_teacher.response": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "uuid.png"
                },
                "birthdate": {
                    "type": "string",
                    "example": "2002-09-09T10:10:10+09:00"
                },
                "common_rate": {
                    "type": "number",
                    "example": 0
                },
                "common_reviews_count": {
                    "type": "integer",
                    "example": 0
                },
                "count_of_students": {
                    "type": "integer",
                    "example": 0
                },
                "email": {
                    "type": "string",
                    "example": "qwerty@example.com"
                },
                "finished_lessons": {
                    "type": "integer",
                    "example": 0
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "registration_date": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10+09:00"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_teacher.skill"
                    }
                },
                "surname": {
                    "type": "string",
                    "example": "Smith"
                },
                "teacher_id": {
                    "type": "integer",
                    "example": 1
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "get_teacher.skill": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string",
                    "example": "about me..."
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_name": {
                    "type": "string",
                    "example": "Category"
                },
                "rate": {
                    "type": "number",
                    "example": 5
                },
                "reviews_count": {
                    "type": "integer",
                    "example": 1
                },
                "skill_id": {
                    "type": "integer",
                    "example": 1
                },
                "video_card_link": {
                    "type": "string",
                    "example": "https://youtu.be/HIcSWuKMwOw?si=FtxN1QJU9ZWnXy85"
                }
            }
        },
...
        "get_teachers.response": {
            "type": "object",
            "properties": {
                "teachers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_teachers.teacher"
                    }
                }
            }
        },
        "get_teachers.skill": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string",
                    "example": "about me..."
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_name": {
                    "type": "string",
                    "example": "Category"
                },
                "rate": {
                    "type": "number",
                    "example": 5
                },
                "reviews_count": {
                    "type": "integer",
                    "example": 1
                },
                "skill_id": {
                    "type": "integer",
                    "example": 1
                },
                "video_card_link": {
                    "type": "string",
                    "example": "https://youtu.be/HIcSWuKMwOw?si=FtxN1QJU9ZWnXy85"
                }
            }
        },
        "get_teachers.teacher": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "uuid.png"
                },
                "birthdate": {
                    "type": "string",
                    "example": "2002-09-09T10:10:10+09:00"
                },
                "common_rate": {
                    "type": "number",
                    "example": 0
                },
                "common_reviews_count": {
                    "type": "integer",
                    "example": 0
                },
                "count_of_students": {
                    "type": "integer",
                    "example": 0
                },
                "email": {
                    "type": "string",
                    "example": "qwerty@example.com"
                },
                "finished_lessons": {
                    "type": "integer",
                    "example": 0
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "registration_date": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10+09:00"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_teachers.skill"
                    }
                },
                "surname": {
                    "type": "string",
                    "example": "Smith"
                },
                "teacher_id": {
                    "type": "integer",
                    "example": 1
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "get_times.response": {
            "type": "object",
            "properties": {
                "datetimes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_times.times"
                    }
                }
            }
        },
        "get_times.times": {
            "type": "object",
            "properties": {
                "datetime": {
                    "type": "string",
                    "example": "0001-01-01T00:00:00Z"
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "schedule_time_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "httputils.ErrorStruct": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}